+++ server/src/main/java/dataaccess/GameSQLDAO.java
@@ -16,10 +16,6 @@ import java.util.HashSet;
 public class GameSQLDAO implements GameDAO {
 
     public GameSQLDAO() {
-        configureDatabase();
-    }
-
-    private void configureDatabase() {
         String DATATABLE_NAME = "gameDatatable";
         try (var conn = DatabaseManager.getConnection()) {
             String tableString = "CREATE table IF NOT EXISTS "+DATATABLE_NAME+" (gameID INT NOT NULL AUTO_INCREMENT PRIMARY KEY, whiteUsername VARCHAR(128), blackUsername VARCHAR(128), gameName VARCHAR(128), chessGame VARCHAR(8192))";
@@ -30,27 +26,18 @@ public class GameSQLDAO implements GameDAO {
         }
     }
 
-
     @Override
     public GameData createGame(GameData gameData) throws DataAccessException {
         try (var conn = DatabaseManager.getConnection()) {
+            if (gameData==null){throw new DataAccessException("passed in null");}
             var statement = "INSERT INTO gameDatatable (whiteUsername, blackUsername, gameName, chessGame) VALUES (?,?,?,?)";
             var preparedStatement = conn.prepareStatement(statement, Statement.RETURN_GENERATED_KEYS);
-//            preparedStatement.setInt(1,gameData.gameID());
             preparedStatement.setString(1,gameData.whiteUsername());
             preparedStatement.setString(2,gameData.blackUsername());
             preparedStatement.setString(3,gameData.gameName());
             String gameAsString = new Gson().toJson(gameData.game());
             preparedStatement.setString(4,gameAsString);
             var id = preparedStatement.executeUpdate();
-//            try (var generatedKeys = preparedStatement.getGeneratedKeys()) {
-//                if (generatedKeys.next()) {
-//                    int gameId = generatedKeys.getInt(1);
-//                    gameData = new GameData(gameId,); // Assuming you have a setter for gameID in GameData
-//                } else {
-//                    throw new SQLException("Creating game failed, no ID obtained.");
-//                }
-//            }
             return gameData;
         }
         catch (SQLException ex) {
@@ -116,8 +103,7 @@ public class GameSQLDAO implements GameDAO {
     public void updateGamePlayer(GameData gameData, ChessGame.TeamColor color, String username) throws DataAccessException {
         try (var conn = DatabaseManager.getConnection()) {
             int gameID = gameData.gameID();
-//            ChessGame myGame = gameData.game();
-//            String gameName = gameData.gameName();
+
             String blackName = gameData.blackUsername();
             String whiteName = gameData.whiteUsername();
 
@@ -157,6 +143,7 @@ public class GameSQLDAO implements GameDAO {
 
     @Override
     public void deleteGame(int gameID) throws DataAccessException {
+        if (gameID<0) {throw new DataAccessException("invalid gameID");}
         try (var conn = DatabaseManager.getConnection()) {
             var statement = "DELETE FROM gameDatatable WHERE gameID = ? " ;
             var preparedStatement = conn.prepareStatement(statement);
